third_party/WebKit/LayoutTests/fast/canvas/canvas-ellipse-negative-radius-expected.txt

Tests CanvasPathMethods ellipse with negative radii.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.ellipse(10, 10, 10, 5, 0, 0, 1, false) did not throw exception.
PASS ctx.ellipse(10, 10, 10, 0, 0, 0, 1, false) did not throw exception.
PASS ctx.ellipse(10, 10, -0, 5, 0, 0, 1, false) did not throw exception.
PASS ctx.ellipse(10, 10, -2, 5, 0, 0, 1, false) threw exception IndexSizeError: Failed to execute 'ellipse' on 'CanvasRenderingContext2D': The major-axis radius provided (-2) is negative..
PASS ctx.ellipse(10, 10, 0, -1.5, 0, 0, 1, false) threw exception IndexSizeError: Failed to execute 'ellipse' on 'CanvasRenderingContext2D': The minor-axis radius provided (-1.5) is negative..
PASS ctx.ellipse(10, 10, -2, -5, 0, 0, 1, false) threw exception IndexSizeError: Failed to execute 'ellipse' on 'CanvasRenderingContext2D': The major-axis radius provided (-2) is negative..
PASS successfullyParsed is true

TEST COMPLETE
